cpufreq: Underclock performance cluster to 1958400 kHz
Underclock the performance cluster from 2457600 kHz to 1958400 kHz for
significant power savings. 1958400 kHz was determined empirically to be
the optimal frequency to underclock to without significantly impacting
performance, given the power consumption of the frequencies directly
above
and below it.

The underclock can be disabled by specifying
"cpufreq.disable_underclock=1"
in the kernel command line, or by writing to disable_underclock at
runtime.
Note that we need to underclock like this instead of outright removing
the
higher frequencies because removing the higher frequencies outright
causes
a significant decrease in performance for some reason.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Michele Beccalossi <beccalossi.michele@gmail.com>