qcacld-3.0: Fix defined but not used compilation error
Commit 75e6f60 ("qcacld-3.0: Load driver on late initcall when not built
as a module") added guards to the portions of code only useful when
building the driver as a module. The recent updates to qcacld-3.0
(as of "LA.UM.6.4.r1-07600-8x98.0") removed a portion of code where
wlan_hdd_state_ctrl_param_destroy was also being used outside of those
guards, leaving it now unused for built-in builds. This made the
compiler trigger a warning (and error) for a declared but unused
variable.

Add guards to the wlan_hdd_state_ctrl_param_destroy function declaration
to avoid declaring it even when not necessary (i.e. not building as a
module).

Signed-off-by: Michele Beccalossi <beccalossi.michele@gmail.com>